ml-service/requirements.txt
# Direct AgriConnect TN ML Service Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2

# Deep Learning (optional)
tensorflow==2.15.0
torch==2.1.1

# Data Processing
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# API
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0

